+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.1.4
 
 - Updated `Highlight` component to have optional highlight prop (#280)
+- Fix `IconLink` without label color style (#357)
 
 ## 2.1.3